# Numba indexing speedups