… benchmarks/benchmark_coo.py
… sparse/coo/indexing.py
… sparse/slicing.py
… sparse/tests/test_coo.py