Significant speedups to indexing. Benchmark:
```python
import sparse
x = sparse.random((1000, 1000, 1000), density=0.01)
%timeit x[5]
```

master:
```
46.9 ms ± 397 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

This branch:
```
174 µs ± 34.5 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Note that this speedup is partially because of Numba but also because our old indexing was completely suboptimal in big-O complexity.

The partial support that we had for advanced indexing is currently broken, but I plan to work on full-blown advanced indexing soon.

As far as I can tell, however, it, too, requires `list[list]]` on Numba's end. xref numba/numba#2560

Fixes #60