@hameerabbasi 2018-03-21
Latest benchmarks:
```
In[2]: import sparse
In[3]: x = sparse.random((1000, 1000, 1000), density=0.01)
In[4]: %timeit x[:500]
78.2 ms ± 613 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
In[5]: %timeit x[:500, :500]
223 ms ± 6.73 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
In[6]: %timeit x[:500, :500, :500]
230 ms ± 6.07 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
In[7]: %timeit x[5]
157 µs ± 423 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Faster in all cases than old. Falls back to simple matching based on a "guesstimate" of which one will be faster.