@hameerabbasi 2018-03-21
cc @mrocklin Ready for review. :-)