@mrocklin 2018-03-21
Cool.  I'm looking forward to it.   I took a brief glance at it and noticed that it was fairly involved, including many new routines with complex logic.  This will probably take some time to review properly and so I may not get to this today.  

If you're inclined I suspect that you could make reviewing (and thus reading by future devs) easier by including small and simple doctests within each utility function that show what it does with a minimal example.