@hameerabbasi 2018-03-21
I added a bit of comments to the main function.