@hameerabbasi 2018-03-24
I think your idea of moving this to a separate module and adding a module-level docstring is probably best. I'll request you to hold off reviewing further until I make that change (it'll make things easier). I'll hopefully do this tomorrow at some point.

The algorithm itself probably isn't that difficult, but I can see the need for terminology being clarified.