@mrocklin 2018-03-24
Thanks.  In general I'm excited about the performance gains here.  Another option would just be to merge and move ahead without review.  Ideally in the future we'll have more people who have time to both write and review code.