@hameerabbasi 2018-03-24
I've clarified everything, hopefully. :-) Feel free to review and/or merge.