@hameerabbasi 2018-03-24
> It seems like a lot of the logic here is based around performance considerations. Do we want to add benchmarks for these various cases?

Seems like an excellent idea!