@mrocklin 2018-03-25
> I'm fairly certain how it will be done, but I haven't mapped it out in code yet. However, it requires list[list[int]] on Numba's end.

How confident are we that this will be implemented soon and, if implemented, that it will be fast enough?  If neither of those are true then do we have a backup plan?