@hameerabbasi 2018-03-25
> How confident are we that this will be implemented soon and, if implemented, that it will be fast enough? If neither of those are true then do we have a backup plan?

numba/numba#2560 is the issue we're looking for. It has a milestone of 0.38 (RC), which is due out April 11. Unless it's pushed to the next release, we should be okay.

There may be a way with flat data structures as well, but it's slightly more complicated. May be more performant, though. I'm shooting for next weekend. It requires some Numba trickery in any case (I think that trickery should be possible in `nopython` mode and if so, it'll be fast, roughly equivalent to `n` integer indexes where `n` is the size of the advanced index).

If nothing works out we can simply loop over the relevant indices in Python.

Do you need it urgently in your work?