@mrocklin 2018-03-25
No, I just want to make sure that we don't over-engineer into a solution that can not be extended.  I suspect that accepting lists and arrays of boolean and integer values will be important for XArray support, which is a nice use case to target.