@mrocklin 2018-03-25
I still haven't gone through all of the logic here, but I've probably gone as far as I'm likely to.  In general things seem fine to me.  +1

Thanks for slogging through this @hameerabbasi .  I hope you're enjoying yourself :)